requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.5.3
Flask-CORS==4.0.0
Flask-Mail==0.9.1
PyMySQL==1.1.0
cryptography==41.0.7
Werkzeug==2.3.7
python-dotenv==1.0.0
Pillow==10.0.1
numpy==1.24.3
gunicorn==21.2.0

# Optional: For actual AI model implementation
# tensorflow==2.13.0
# or
# torch==2.0.1
# torchvision==0.15.2